feat: build multiple books

+++ packages/application/src/file-operations/build-operations/epub-build.ts
@@ -1,21 +1,20 @@
-import { DISTROBUTION_FOLDER_PATH } from "~file-operations";
+import { DISTROBUTION_FOLDER_PATH } from "~constants";
 import { runCommand } from "~helpers";
 import logger from "~logger";
-import { BookConfigs } from "~types";
 
-export async function epubBuild(bookConfigs: BookConfigs): Promise<void> {
-  for (const bookConfig of bookConfigs) {
-    const { epub, name } = bookConfig;
-    const ebook = `${name}.epub`;
-    const distrobutionFile = `../../../${DISTROBUTION_FOLDER_PATH}/${name}.epub`;
+export async function epubBuild(path: string, name: string): Promise<void> {
+  if (!path || !name) {
+    throw new Error("The path and name are required to build the epub file.");
+  }
+  const ebook = `${name}.epub`;
+  const distrobutionFile = `../../../${DISTROBUTION_FOLDER_PATH}/${name}.epub`;
 
-    await runCommand(`
-      cd ${epub} && 
+  await runCommand(`
+      cd ${path} && 
       rm -f ${distrobutionFile} &&
       zip -X0 ./${ebook} mimetype && 
       zip -r ${ebook} META-INF epub && 
       mv ${ebook} ${distrobutionFile}`);
 
-    logger.notice(`The '${ebook}' file was created successfully.`);
-  }
+  logger.notice(`The '${ebook}' file was created successfully.`);
 }
